# Cache Jinja environment by response type