--- a/tests/test_core/test_responses.py
+++ b/tests/test_core/test_responses.py
@@ -163,3 +163,41 @@ def test_get_dict_list_params():
     result = subject._get_multi_param_dict("VpcSecurityGroupIds")
 
     result.should.equal({"VpcSecurityGroupId": ["sg-123", "sg-456", "sg-789"]})
+
+
+def test_response_environment_preserved_by_type():
+    """Ensure Jinja environment is cached by response type."""
+
+    class ResponseA(BaseResponse):
+        pass
+
+    class ResponseB(BaseResponse):
+        pass
+
+    resp_a = ResponseA()
+    another_resp_a = ResponseA()
+    resp_b = ResponseB()
+
+    assert resp_a.environment is another_resp_a.environment
+    assert resp_b.environment is not resp_a.environment
+
+    source_1 = "template"
+    source_2 = "amother template"
+
+    assert not resp_a.contains_template(id(source_1))
+    resp_a.response_template(source_1)
+    assert resp_a.contains_template(id(source_1))
+
+    assert not resp_a.contains_template(id(source_2))
+    resp_a.response_template(source_2)
+    assert resp_a.contains_template(id(source_2))
+
+    assert not resp_b.contains_template(id(source_1))
+    assert not resp_b.contains_template(id(source_2))
+
+    assert another_resp_a.contains_template(id(source_1))
+    assert another_resp_a.contains_template(id(source_2))
+
+    resp_a_new_instance = ResponseA()
+    assert resp_a_new_instance.contains_template(id(source_1))
+    assert resp_a_new_instance.contains_template(id(source_2))